move unsafeFind to extended ORD_MAP
This is some slightly useless module-level wankery that's needless to
say kind of cool.

unsafeFind: 'a map * Key.ord_key -> 'a is a useful little function
that makes it clear that something could go wrong (exception Option,
although there may be a choice better exception)

It (and any other functions you want) can be added to an extended
version of ORD_MAP using a functor; that's MapUtilsFn.